update: 기존에 바이트로 넘어온 값을 메가바이트 수정

+++ src/main/java/com/develop/web/domain/service/clip/dto/ClipDto.java
@@ -28,7 +28,7 @@ public class ClipDto {
     public String format_long_name;
     public String tags;
     public double duration;
-    public Long size;
+    public double size;
 
     public Date ingest_at;
     public LocalDateTime end_at;
